requirements.txt
streamlit
google-generativeai
pandas
pillow
sqlite3
